Refactor srceener into a dedicated package
