+++ apps/parfait-list/main_test.go
@@ -29,7 +29,7 @@ import (
 func TestMain(t *testing.T) {
 	t.Parallel()
 
-	tmpdir, tmpdirErr := os.MkdirTemp("", "test_list")
+	tmpdir, tmpdirErr := os.MkdirTemp("", "test_list_app")
 	defer os.RemoveAll(tmpdir)
 
 	Convey("Setup succeeded", t, func() {
